Move cleanup to its own fail
